refactor: Store setState api 반환하지 않도록 수정
