+++ src/main/kotlin/racingcar/store/Store.kt
@@ -2,5 +2,5 @@ package racingcar.store
 
 interface Store<T> {
     fun getState(): T
-    fun setState(state: T): T
+    fun setState(state: T)
 }
